security-manager: fix build issues for dizzy
With version dizzy of poky, the library libcap doesn't
provide file for pkgconfig.

Without this patch the build fails because, first,
cmake complains that libcap isn't installed, and,
second, the extended attribute names for smack are
not found.

This commits insert a check that decide if the 2 added
patches that solve these issues have to be applied or not.

Change-Id: I6e18373fd2253dcd70f17a6da773049010fc3e54
Signed-off-by: José Bollo <jose.bollo@iot.bzh>