dbus-cynara_1.8.18.bb: ignore patch for 1.8.2
CVE-2015-0245-prevent-forged-ActivationFailure.patch was added to
dbus.inc in dizzy and fido for use with D-Bus 1.8.2, but is not
needed for our version of D-Bus because it already contains that
fix.

Perhaps it would have been better to add it upstream to dbus_1.8.2.bb
instead of dbus.inc, because it is not reusable across different D-Bus
versions.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>