application_privilege.py: fix copying of notroot.py
The fallback code for installing notroot.py in /tmp when it is not
already present was broken: notroot.py does not get deployed and thus
is not found in the directory pointed to by get_files_dir(). We have to
copy it directly from the meta-security-smack files folder.

The assumption here is that the two layers are located side-by-side, as in
the meta-intel-iot-security repository. Using layers from two different places
will not work as expected.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>